pyproject.toml
[project]
name = "shyn-note-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.115.11",
    "psycopg2-binary>=2.9.10",
    "ruff>=0.9.10",
    "uvicorn>=0.34.0",
]
